chore(ios): remove dead LibraryFilterTabs (fixes #1006)
The Library type filter moved to a pull-down dropdown (LibraryFilterMenu), so
`struct LibraryFilterTabs` was dead — only its own preview and two snapshot
tests referenced it. Extract `enum LibraryFilter` (used app-wide) into
LibraryFilter.swift, delete the struct + its two snapshot tests + orphaned PNGs,
and fix the two doc comments that named it. The inline-scrolling SegmentedPills
layout is now unused but retained as a reusable primitive variant.

Fixes #1006
